FFT printing works
The compute_fft() function needs to be split up and optimized. Next step is using struct to have all the "plotting" information one place so we can also display axis and potentially labels etc.

Will also try to explore a zoom option, but a bit unsure how to get that to work without a cursor to show where you are on the LCD.